Add before_run and after_run logging
