+++ griptape/tasks/audio_transcription_task.py
@@ -1,36 +1,17 @@
 from __future__ import annotations
 
-from abc import ABC
-from typing import Callable
-
 from attrs import define, field
 
-from griptape.artifacts.audio_artifact import AudioArtifact
 from griptape.engines import AudioTranscriptionEngine
 from griptape.artifacts import TextArtifact
-from griptape.mixins import RuleMixin
-from griptape.tasks import BaseTask
+from griptape.tasks.base_audio_input_task import BaseAudioInputTask
 
 
 @define
-class AudioTranscriptionTask(RuleMixin, BaseTask, ABC):
-    _input: AudioArtifact | Callable[[BaseTask], AudioArtifact] = field(alias="input")
+class AudioTranscriptionTask(BaseAudioInputTask):
     _audio_transcription_engine: AudioTranscriptionEngine = field(
         default=None, kw_only=True, alias="audio_transcription_engine"
     )
-
-    @property
-    def input(self) -> AudioArtifact:
-        if isinstance(self._input, AudioArtifact):
-            return self._input
-        elif isinstance(self._input, Callable):
-            return self._input(self)
-        else:
-            raise ValueError("Input must be an AudioArtifact.")
-
-    @input.setter
-    def input(self, value: AudioArtifact | Callable[[BaseTask], AudioArtifact]) -> None:
-        self._input = value
 
     @property
     def audio_transcription_engine(self) -> AudioTranscriptionEngine:
@@ -46,16 +27,6 @@ class AudioTranscriptionTask(RuleMixin, BaseTask, ABC):
     @audio_transcription_engine.setter
     def audio_transcription_engine(self, value: AudioTranscriptionEngine) -> None:
         self._audio_transcription_engine = value
-
-    def before_run(self) -> None:
-        super().before_run()
-
-        self.structure.logger.info(f"{self.__class__.__name__} {self.id}\nInput: {self.input.to_text()}")
-
-    def after_run(self) -> None:
-        super().after_run()
-
-        self.structure.logger.info(f"{self.__class__.__name__} {self.id}\nOutput: {self.output.to_text()}")
 
     def run(self) -> TextArtifact:
         return self.audio_transcription_engine.run(self.input)
